Script to touch all files
Use locally then push to Git when you want Weblate to re-read all
translation files. The script works by dicking around with
newlines at the end of the file (adding or removing one as
needed). It does not use a simple touch() because on Windows that
does nothing that Git will pick up. Yeah, I know...